+++ cassandra/cluster.py
@@ -852,8 +852,8 @@ class Cluster(object):
                                       "statement on host %s: %r", host, response)
 
             log.debug("Done preparing all known prepared statements against host %s", host)
-        except OperationTimedOut:
-            log.warn("Timed out trying to prepare all statements on host %s", host)
+        except OperationTimedOut as timeout:
+            log.warn("Timed out trying to prepare all statements on host %s: %s", host, timeout)
         except (ConnectionException, socket.error) as exc:
             log.warn("Error trying to prepare all statements on host %s: %r", host, exc)
         except Exception:
@@ -1644,8 +1644,9 @@ class ControlConnection(object):
                     timeout = min(2.0, total_timeout - elapsed)
                     peers_result, local_result = connection.wait_for_responses(
                         peers_query, local_query, timeout=timeout)
-                except OperationTimedOut:
-                    log.debug("[control connection] Timed out waiting for response during schema agreement check")
+                except OperationTimedOut as timeout:
+                    log.debug("[control connection] Timed out waiting for " \
+                              "response during schema agreement check: %s", timeout)
                     elapsed = self._time.time() - start
                     continue
 
@@ -2189,7 +2190,7 @@ class ResponseFuture(object):
             elif self._final_exception:
                 raise self._final_exception
             else:
-                raise OperationTimedOut()
+                raise OperationTimedOut(errors=self._errors, last_host=self._current_host)
 
     def get_query_trace(self, max_wait=None):
         """